pyproject.toml
[project]
name = "fastapi-mini-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "tortoise-orm[asyncpg]>=0.25.3",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
prod = [
    "aerich>=0.9.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "tortoise-orm>=0.25.3",
    "uvicorn[standard]>=0.40.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]  # 이 섹션 아래로 옮겨야 경고가 사라집니다
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.aerich]
tortoise_config = "app.db.base.TORTOISE_CONFIG"
location = "app/db/migrations"
src_folder = "./"
